# Use strings as keys until both non-string ids are used everywhere